Refactor JsonSchema

+++ Maven-AT/src/main/java/models/order/Order.java
@@ -7,7 +7,7 @@ import utils.base.Model;
 @Data
 @Builder
 public class Order extends Model<Order> {
-	private int petId;
+	private Long petId;
 	private int quantity;
 	private int id;
 	private String shipDate;
